+++ elastalert/loaders.py
@@ -25,6 +25,7 @@ import elastalert.alerters.gelf
 import elastalert.alerters.googlechat
 import elastalert.alerters.httppost
 import elastalert.alerters.httppost2
+import elastalert.alerters.iris
 import elastalert.alerters.lark
 import elastalert.alerters.line
 import elastalert.alerters.pagertree
@@ -132,7 +133,8 @@ class RulesLoader(object):
         'datadog': elastalert.alerters.datadog.DatadogAlerter,
         'ses': elastalert.alerters.ses.SesAlerter,
         'rocketchat': elastalert.alerters.rocketchat.RocketChatAlerter,
-        'gelf': elastalert.alerters.gelf.GelfAlerter
+        'gelf': elastalert.alerters.gelf.GelfAlerter,
+        'iris': elastalert.alerters.iris.IrisAlerter,
     }
 
     # A partial ordering of alert types. Relative order will be preserved in the resulting alerts list